pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-assistant"
version = "0.1.0"
description = "Kernel + domain-plugin second-brain agent backed by claude -p, Telegram, and an Obsidian/Drive vault."
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Jason Chi" },
]
dependencies = [
  "python-telegram-bot ~= 21.0",
  "pyyaml >= 6.0, < 7.0",
  "pydantic ~= 2.0",
  "tenacity >= 8.0, < 10.0",
  "python-dateutil >= 2.8, < 3.0",
  "rich >= 13.0, < 15.0",
]

[project.optional-dependencies]
dev = [
  "pytest >= 8.0, < 9.0",
  "pytest-mock >= 3.10, < 4.0",
]

[tool.setuptools.packages.find]
include = ["kernel*"]
exclude = ["tests*", "vault*", "domains*", "eval*", "configs*", "plans*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-ra"
